user/requirements.txt
fastapi[standard]
sqlalchemy[asyncio] 
alembic
asyncpg
psycopg2-binary
bcrypt
python-decouple